remove code smells
