+++ src/app/store/state/ecoEvents.state.ts
@@ -1,4 +1,4 @@
-import { EventPageResponceDto, EventResponseDto } from 'src/app/main/component/events/models/events.interface';
+import { EventResponseDto } from 'src/app/main/component/events/models/events.interface';
 
 export interface IEcoEventsState {
   eventState: EventResponseDto;
