--- src/app/ubs/ubs-admin/components/shared/components/dialog-pop-up/dialog-pop-up.component.html
@@ -1,10 +1,0 @@
-<div class="main-container">
-  <div class="warning-text">
-    <div class="warning-title" *ngIf="popupTitle">{{ popupTitle | translate }}</div>
-    <div class="warning-subtitle" *ngIf="popupSubtitle">{{ popupSubtitle | translate }}</div>
-  </div>
-  <div class="buttons-container">
-    <button class="secondary-global-button" (click)="userReply(false)">{{ popupCancel | translate }}</button>
-    <button class="primary-global-button" (click)="userReply(true)">{{ popupConfirm | translate }}</button>
-  </div>
-</div>
