--- src/app/ubs/ubs-admin/components/shared/components/dialog-pop-up/dialog-pop-up.component.ts
@@ -1,56 +1,0 @@
-import { Component, Inject, OnDestroy, OnInit } from '@angular/core';
-import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { Subject } from 'rxjs';
-import { takeUntil } from 'rxjs/operators';
-
-@Component({
-  selector: 'app-dialog-pop-up',
-  templateUrl: './dialog-pop-up.component.html',
-  styleUrls: ['./dialog-pop-up.component.scss']
-})
-export class DialogPopUpComponent implements OnInit, OnDestroy {
-  private destroy$: Subject<boolean> = new Subject<boolean>();
-  popupTitle: string;
-  popupSubtitle: string;
-  popupConfirm: string;
-  popupCancel: string;
-
-  constructor(private matDialogRef: MatDialogRef<DialogPopUpComponent>, @Inject(MAT_DIALOG_DATA) public data) {}
-
-  ngOnInit(): void {
-    this.setTitles();
-    this.matDialogRef
-      .keydownEvents()
-      .pipe(takeUntil(this.destroy$))
-      .subscribe((event) => {
-        if (event.key === 'Escape') {
-          this.userReply(false);
-        }
-        if (event.key === 'Enter') {
-          this.userReply(true);
-        }
-      });
-    this.matDialogRef
-      .backdropClick()
-      .pipe(takeUntil(this.destroy$))
-      .subscribe(() => {
-        this.userReply(false);
-      });
-  }
-
-  private setTitles(): void {
-    this.popupTitle = this.data.popupTitle;
-    this.popupSubtitle = this.data.popupSubtitle;
-    this.popupConfirm = this.data.popupConfirm;
-    this.popupCancel = this.data.popupCancel;
-  }
-
-  public userReply(reply: boolean): void {
-    this.matDialogRef.close(reply);
-  }
-
-  ngOnDestroy(): void {
-    this.destroy$.next();
-    this.destroy$.complete();
-  }
-}
